# src/routes/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Record_string.unknown_": {
				"properties": {},
				"additionalProperties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
...
			"RecordId": {
				"type": "string",
				"example": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e"
			},
...
			"ProofExchangeRecord": {
				"$ref": "#/components/schemas/Record_string.unknown_"
			},
...
			"AcceptProofProposal": {
				"properties": {
					"proofRecordId": {
						"type": "string"
					},
					"proofFormats": {},
					"comment": {
						"type": "string"
					},
					"autoAcceptProof": {
						"$ref": "#/components/schemas/AutoAcceptProof"
					},
					"goalCode": {
						"type": "string"
					},
					"willConfirm": {
						"type": "boolean"
					}
				},
				"required": [
					"proofRecordId",
					"proofFormats"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"PlaintextMessage": {
				"properties": {
					"@type": {
						"type": "string"
					},
					"@id": {
						"type": "string"
					},
					"~thread": {
						"properties": {
							"pthid": {
								"type": "string"
							},
							"thid": {
								"type": "string"
							}
						},
						"type": "object"
					}
				},
				"required": [
					"@type",
					"@id"
				],
				"type": "object",
				"additionalProperties": {}
			},
...
			"RecipientKeyOption": {
				"properties": {
					"recipientKey": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"ThreadId": {
				"type": "string",
				"example": "ea4e5e69-fc04-465a-90d2-9f8ff78aa71d"
			},
...
		},
		"securitySchemes": {
			"apiKey": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			},
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "credo-controller",
		"version": "2.0.0",
		"description": "Rest endpoint wrapper for using your agent over HTTP",
		"license": {
			"name": "Apache-2.0"
		},
		"contact": {}
	},
	"paths": {
...
		"/didcomm/proofs/{proofRecordId}/accept-proposal": {
			"post": {
				"operationId": "AcceptProposal",
				"responses": {
					"200": {
						"description": "ProofRecord",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProofExchangeRecord"
								},
								"examples": {
									"Example 1": {
										"value": {
											"metadata": {},
											"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
											"createdAt": "2022-01-01T00:00:00.000Z",
											"connectionId": "2aecf74c-3073-4f98-9acb-92415d096834",
											"threadId": "0019d466-5eea-4269-8c40-031b4896c5b7",
											"protocolVersion": "v1"
										}
									}
								}
							}
						}
					}
				},
				"description": "Accept a presentation proposal as verifier by sending an accept proposal message\nto the connection associated with the proof record.",
				"tags": [
					"DIDComm - Proofs"
				],
				"security": [
					{
						"jwt": [
							"tenant",
							"dedicated"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AcceptProofProposal"
							}
						}
					}
				}
			}
		},
...
		"/didcomm/proofs/{proofRecordId}/accept-request": {
			"post": {
				"operationId": "AcceptRequest",
				"responses": {
					"200": {
						"description": "ProofRecord",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Record_string.unknown_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"metadata": {},
											"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
											"createdAt": "2022-01-01T00:00:00.000Z",
											"connectionId": "2aecf74c-3073-4f98-9acb-92415d096834",
											"threadId": "0019d466-5eea-4269-8c40-031b4896c5b7",
											"protocolVersion": "v1"
										}
									}
								}
							}
						}
					}
				},
				"description": "Accept a presentation request as prover by sending an accept request message\nto the connection associated with the proof record.",
				"tags": [
					"DIDComm - Proofs"
				],
				"security": [
					{
						"jwt": [
							"tenant",
							"dedicated"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "proofRecordId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"comment": {
										"type": "string"
									},
									"filterByNonRevocationRequirements": {
										"type": "boolean"
									},
									"filterByPresentationPreview": {
										"type": "boolean"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
...
		"/didcomm/oob/create-invitation": {
			"post": {
				"operationId": "CreateInvitation",
				"responses": {
					"200": {
						"description": "Out of band record",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"invitationDid": {
											"type": "string"
										},
										"outOfBandRecord": {
											"$ref": "#/components/schemas/Record_string.unknown_"
										},
										"invitation": {
											"$ref": "#/components/schemas/PlaintextMessage"
										},
										"invitationUrl": {
											"type": "string"
										}
									},
									"required": [
										"invitationDid",
										"outOfBandRecord",
										"invitation",
										"invitationUrl"
									],
									"type": "object"
								},
								"examples": {
									"Example 1": {
										"value": {
											"invitationUrl": "string",
											"invitation": {
												"@type": "did:sov:BzCbsNYhMrjHiqZDTUASHg;spec/out-of-band/1.1/invitation",
												"@id": "d6472943-e5d0-4d95-8b48-790ed5a41931",
												"label": "Aries Test Agent",
												"accept": [
													"didcomm/aip1",
													"didcomm/aip2;env=rfc19"
												],
												"handshake_protocols": [
													"https://didcomm.org/didexchange/1.0",
													"https://didcomm.org/connections/1.0"
												],
												"services": [
													{
														"id": "#inline-0",
														"serviceEndpoint": "https://6b77-89-20-162-146.ngrok.io",
														"type": "did-communication",
														"recipientKeys": [
															"did:key:z6MkmTBHTWrvLPN8pBmUj7Ye5ww9GiacXCYMNVvpScSpf1DM"
														],
														"routingKeys": []
													}
												]
											},
											"outOfBandRecord": {
												"_tags": {
													"invitationId": "1cbd22e4-1906-41e9-8807-83d84437f978",
													"state": "await-response",
													"role": "sender",
													"recipientKeyFingerprints": [
														"z6MktUCPZjfRJXD4GMcYuXiqX2qZ8vBw6UAYpDFiHEUfwuLj"
													]
												},
												"outOfBandInvitation": {
													"@type": "did:sov:BzCbsNYhMrjHiqZDTUASHg;spec/out-of-band/1.1/invitation",
													"@id": "d6472943-e5d0-4d95-8b48-790ed5a41931",
													"label": "Aries Test Agent",
													"accept": [
														"didcomm/aip1",
														"didcomm/aip2;env=rfc19"
													],
													"handshake_protocols": [
														"https://didcomm.org/didexchange/1.0",
														"https://didcomm.org/connections/1.0"
													],
													"services": [
														{
															"id": "#inline-0",
															"serviceEndpoint": "https://6b77-89-20-162-146.ngrok.io",
															"type": "did-communication",
															"recipientKeys": [
																"did:key:z6MkmTBHTWrvLPN8pBmUj7Ye5ww9GiacXCYMNVvpScSpf1DM"
															],
															"routingKeys": []
														}
													]
												},
												"metadata": {},
												"id": "42a95528-0e30-4f86-a462-0efb02178b53",
												"createdAt": "2022-01-01T00:00:00.000Z",
												"reusable": false
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Creates an outbound out-of-band record containing out-of-band invitation message defined in\nAries RFC 0434: Out-of-Band Protocol 1.1.",
				"tags": [
					"DIDComm - Out Of Band"
				],
				"security": [
					{
						"jwt": [
							"tenant",
							"dedicated"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "configuration of how out-of-band invitation should be created",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"allOf": [
									{
										"$ref": "#/components/schemas/CreateInvitationOptions"
									},
									{
										"$ref": "#/components/schemas/RecipientKeyOption"
									}
								],
								"description": "configuration of how out-of-band invitation should be created"
							}
						}
					}
				}
			}
		},
...
		"/didcomm/credentials": {
			"get": {
				"operationId": "GetAllCredentials",
				"responses": {
					"200": {
						"description": "CredentialExchangeRecord[]",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Record_string.unknown_"
									},
									"type": "array"
								},
								"examples": {
									"Example 1": {
										"value": [
											{
												"_tags": {
													"state": "offer-sent",
													"threadId": "82701488-b43c-4d7b-9244-4bb204a7ae26",
													"connectionId": "ac6d0fdd-0db8-4f52-8a3d-de7ff8ddc14b"
												},
												"metadata": {
													"_internal/indyCredential": {
														"credentialDefinitionId": "q7ATwTYbQDgiigVijUAej:3:CL:318187:latest",
														"schemaId": "q7ATwTYbQDgiigVijUAej:2:Employee Badge:1.0"
													}
												},
												"credentials": [],
												"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
												"createdAt": "2022-01-01T00:00:00.000Z",
												"connectionId": "ac6d0fdd-0db8-4f52-8a3d-de7ff8ddc14b",
												"threadId": "82701488-b43c-4d7b-9244-4bb204a7ae26",
												"credentialAttributes": [],
												"protocolVersion": "v1"
											}
										]
									}
								}
							}
						}
					}
				},
				"description": "Retrieve all credential exchange records",
				"tags": [
					"DIDComm - Credentials"
				],
				"security": [
					{
						"jwt": [
							"tenant",
							"dedicated"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "threadId",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/ThreadId"
						}
					},
					{
						"in": "query",
						"name": "parentThreadId",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/ThreadId"
						}
					},
					{
						"in": "query",
						"name": "connectionId",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					},
					{
						"in": "query",
						"name": "state",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/CredentialState"
						}
					},
					{
						"in": "query",
						"name": "role",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/CredentialRole"
						}
					}
				]
			}
		},
		"/didcomm/credentials/w3c": {
			"get": {
				"operationId": "GetAllW3c",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/W3cCredentialRecord"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"DIDComm - Credentials"
				],
				"security": [
					{
						"jwt": [
							"tenant",
							"dedicated"
						]
					}
				],
				"parameters": []
			}
		},
		"/didcomm/credentials/w3c/{id}": {
			"get": {
				"operationId": "GetW3cById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/W3cCredentialRecord"
								}
							}
						}
					}
				},
				"tags": [
					"DIDComm - Credentials"
				],
				"security": [
					{
						"jwt": [
							"tenant",
							"dedicated"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/didcomm/credentials/{credentialRecordId}": {
			"get": {
				"operationId": "GetCredentialById",
				"responses": {
					"200": {
						"description": "CredentialExchangeRecord",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Record_string.unknown_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"_tags": {
												"state": "offer-sent",
												"threadId": "82701488-b43c-4d7b-9244-4bb204a7ae26",
												"connectionId": "ac6d0fdd-0db8-4f52-8a3d-de7ff8ddc14b"
											},
											"metadata": {
												"_internal/indyCredential": {
													"credentialDefinitionId": "q7ATwTYbQDgiigVijUAej:3:CL:318187:latest",
													"schemaId": "q7ATwTYbQDgiigVijUAej:2:Employee Badge:1.0"
												}
											},
											"credentials": [],
											"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
											"createdAt": "2022-01-01T00:00:00.000Z",
											"connectionId": "ac6d0fdd-0db8-4f52-8a3d-de7ff8ddc14b",
											"threadId": "82701488-b43c-4d7b-9244-4bb204a7ae26",
											"credentialAttributes": [],
											"protocolVersion": "v1"
										}
									}
								}
							}
						}
					}
				},
				"description": "Retrieve credential exchange record by credential record id",
				"tags": [
					"DIDComm - Credentials"
				],
				"security": [
					{
						"jwt": [
							"tenant",
							"dedicated"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "credentialRecordId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					}
				]
			}
		},
		"/didcomm/credentials/propose-credential": {
			"post": {
				"operationId": "ProposeCredential",
				"responses": {
					"200": {
						"description": "CredentialExchangeRecord",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CredentialExchangeRecord"
								},
								"examples": {
									"Example 1": {
										"value": {
											"_tags": {
												"state": "offer-sent",
												"threadId": "82701488-b43c-4d7b-9244-4bb204a7ae26",
												"connectionId": "ac6d0fdd-0db8-4f52-8a3d-de7ff8ddc14b"
											},
											"metadata": {
												"_internal/indyCredential": {
													"credentialDefinitionId": "q7ATwTYbQDgiigVijUAej:3:CL:318187:latest",
													"schemaId": "q7ATwTYbQDgiigVijUAej:2:Employee Badge:1.0"
												}
											},
											"credentials": [],
											"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
											"createdAt": "2022-01-01T00:00:00.000Z",
											"connectionId": "ac6d0fdd-0db8-4f52-8a3d-de7ff8ddc14b",
											"threadId": "82701488-b43c-4d7b-9244-4bb204a7ae26",
											"credentialAttributes": [],
											"protocolVersion": "v1"
										}
									}
								}
							}
						}
					}
				},
				"description": "Initiate a new credential exchange as holder by sending a propose credential message\nto the connection with a specified connection id.",
				"tags": [
					"DIDComm - Credentials"
				],
				"security": [
					{
						"jwt": [
							"tenant",
							"dedicated"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProposeCredentialOptions"
							}
						}
					}
				}
			}
		},
		"/didcomm/credentials/accept-proposal": {
			"post": {
				"operationId": "AcceptProposal",
				"responses": {
					"200": {
						"description": "CredentialExchangeRecord",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CredentialExchangeRecord"
								},
								"examples": {
									"Example 1": {
										"value": {
											"_tags": {
												"state": "offer-sent",
												"threadId": "82701488-b43c-4d7b-9244-4bb204a7ae26",
												"connectionId": "ac6d0fdd-0db8-4f52-8a3d-de7ff8ddc14b"
											},
											"metadata": {
												"_internal/indyCredential": {
													"credentialDefinitionId": "q7ATwTYbQDgiigVijUAej:3:CL:318187:latest",
													"schemaId": "q7ATwTYbQDgiigVijUAej:2:Employee Badge:1.0"
												}
											},
											"credentials": [],
											"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
											"createdAt": "2022-01-01T00:00:00.000Z",
											"connectionId": "ac6d0fdd-0db8-4f52-8a3d-de7ff8ddc14b",
											"threadId": "82701488-b43c-4d7b-9244-4bb204a7ae26",
											"credentialAttributes": [],
											"protocolVersion": "v1"
										}
									}
								}
							}
						}
					}
				},
				"description": "Accept a credential proposal as issuer by sending an accept proposal message\nto the connection associated with the credential exchange record.",
				"tags": [
					"DIDComm - Credentials"
				],
				"security": [
					{
						"jwt": [
							"tenant",
							"dedicated"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AcceptCredentialProposalOptions"
							}
						}
					}
				}
			}
		},
		"/didcomm/credentials/create-offer": {
			"post": {
				"operationId": "CreateOffer",
				"responses": {
					"200": {
						"description": "AgentMessage, CredentialExchangeRecord",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CredentialExchangeRecord"
								},
								"examples": {
									"Example 1": {
										"value": {
											"_tags": {
												"state": "offer-sent",
												"threadId": "82701488-b43c-4d7b-9244-4bb204a7ae26",
												"connectionId": "ac6d0fdd-0db8-4f52-8a3d-de7ff8ddc14b"
											},
											"metadata": {
												"_internal/indyCredential": {
													"credentialDefinitionId": "q7ATwTYbQDgiigVijUAej:3:CL:318187:latest",
													"schemaId": "q7ATwTYbQDgiigVijUAej:2:Employee Badge:1.0"
												}
											},
											"credentials": [],
											"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
											"createdAt": "2022-01-01T00:00:00.000Z",
											"connectionId": "ac6d0fdd-0db8-4f52-8a3d-de7ff8ddc14b",
											"threadId": "82701488-b43c-4d7b-9244-4bb204a7ae26",
											"credentialAttributes": [],
											"protocolVersion": "v1"
										}
									}
								}
							}
						}
					}
				},
				"description": "Initiate a new credential exchange as issuer by creating a credential offer\nwithout specifying a connection id",
				"tags": [
					"DIDComm - Credentials"
				],
				"security": [
					{
						"jwt": [
							"tenant",
							"dedicated"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateOfferOptions"
							}
						}
					}
				}
			}
		},
		"/didcomm/credentials/create-offer-oob": {
			"post": {
				"operationId": "CreateOfferOob",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"invitationDid": {
											"type": "string"
										},
										"credentialRequestThId": {
											"type": "string"
										},
										"outOfBandRecordId": {
											"type": "string"
										},
										"outOfBandRecord": {
											"$ref": "#/components/schemas/Record_string.unknown_"
										},
										"invitation": {
											"$ref": "#/components/schemas/PlaintextMessage"
										},
										"invitationUrl": {
											"type": "string"
										}
									},
									"required": [
										"invitationDid",
										"credentialRequestThId",
										"outOfBandRecordId",
										"outOfBandRecord",
										"invitation",
										"invitationUrl"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"DIDComm - Credentials"
				],
				"security": [
					{
						"jwt": [
							"tenant",
							"dedicated"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateOfferOobOptions"
							}
						}
					}
				}
			}
		},
		"/didcomm/credentials/accept-offer": {
			"post": {
				"operationId": "AcceptOffer",
				"responses": {
					"200": {
						"description": "CredentialExchangeRecord",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CredentialExchangeRecord"
								},
								"examples": {
									"Example 1": {
										"value": {
											"_tags": {
												"state": "offer-sent",
												"threadId": "82701488-b43c-4d7b-9244-4bb204a7ae26",
												"connectionId": "ac6d0fdd-0db8-4f52-8a3d-de7ff8ddc14b"
											},
											"metadata": {
												"_internal/indyCredential": {
													"credentialDefinitionId": "q7ATwTYbQDgiigVijUAej:3:CL:318187:latest",
													"schemaId": "q7ATwTYbQDgiigVijUAej:2:Employee Badge:1.0"
												}
											},
											"credentials": [],
											"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
											"createdAt": "2022-01-01T00:00:00.000Z",
											"connectionId": "ac6d0fdd-0db8-4f52-8a3d-de7ff8ddc14b",
											"threadId": "82701488-b43c-4d7b-9244-4bb204a7ae26",
											"credentialAttributes": [],
											"protocolVersion": "v1"
										}
									}
								}
							}
						}
					}
				},
				"description": "Accept a credential offer as holder by sending an accept offer message\nto the connection associated with the credential exchange record.",
				"tags": [
					"DIDComm - Credentials"
				],
				"security": [
					{
						"jwt": [
							"tenant",
							"dedicated"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CredentialOfferOptions"
							}
						}
					}
				}
			}
		},
		"/didcomm/credentials/accept-request": {
			"post": {
				"operationId": "AcceptRequest",
				"responses": {
					"200": {
						"description": "CredentialExchangeRecord",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CredentialExchangeRecord"
								},
								"examples": {
									"Example 1": {
										"value": {
											"_tags": {
												"state": "offer-sent",
												"threadId": "82701488-b43c-4d7b-9244-4bb204a7ae26",
												"connectionId": "ac6d0fdd-0db8-4f52-8a3d-de7ff8ddc14b"
											},
											"metadata": {
												"_internal/indyCredential": {
													"credentialDefinitionId": "q7ATwTYbQDgiigVijUAej:3:CL:318187:latest",
													"schemaId": "q7ATwTYbQDgiigVijUAej:2:Employee Badge:1.0"
												}
											},
											"credentials": [],
											"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
											"createdAt": "2022-01-01T00:00:00.000Z",
											"connectionId": "ac6d0fdd-0db8-4f52-8a3d-de7ff8ddc14b",
											"threadId": "82701488-b43c-4d7b-9244-4bb204a7ae26",
											"credentialAttributes": [],
											"protocolVersion": "v1"
										}
									}
								}
							}
						}
					}
				},
				"description": "Accept a credential request as issuer by sending an accept request message\nto the connection associated with the credential exchange record.",
				"tags": [
					"DIDComm - Credentials"
				],
				"security": [
					{
						"jwt": [
							"tenant",
							"dedicated"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AcceptCredentialRequestOptions"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}